mcp/requirements.txt
fastmcp>=2.12.0
uvicorn>=0.35.0
browser-use>=0.7.7
anthropic>=0.58.2
python-dotenv>=1.0.0
